Add files via upload
this css file has the styling for the header, the navigation or filter bar, the option buttons in the filter bar, the main container, the cards, the placement of the user's image and details inside the cards, and the styling of the user input bar at the bottom. Font has been imported from google fonts and images have been used which will be attached separately. In the filter bar, the buttons have an on click function but the icons do not ( I had a tough time trying to align everything )